Cascade Terminate and Purge Support (#47)
This change fixes issues in current implementation of Cascade Terminate and adds support for cascade Purge. It is required to support dapr/dapr#6393. As per current implementation,

Each language SDK must implement the cascade terminate feature

A failed sub-orchestration in the chain might block termination of its own downstream sub-orchestrations

This PR moves the recursive termination logic to backend side and so no SDK would need to have any implementation. Also it fixes the issue of failed orchestration blocking termination of its downstream sub-orchestrations. It also adds a new test to validate both scenarios. (The second test added failed for the original implementation.)

This PR also implements Cascade Purge, similar to Cascade terminate and adds tests for the same. The Proto updates are corresponding to the protobuf updates: microsoft/durabletask-protobuf#20

Signed-off-by: Shivam Kumar <shivamkm07@gmail.com>