Adds RerunWorkflowFromEvento
Adds new gRPC server and client methods to implement reruning a
workflow from a given event ID.

The client can optionally give a new input to the event which the
workflow is being rerun from. Because of the improper choice of
`StringValue` for the proto input field, we require a `bool` signal flag
to allow for setting `nil` values.

Can optionally give a new instance ID to create, or one will be
generated for the client.

Signed-off-by: joshvanl <me@joshvanl.dev>