Support loadbalance of durable task clients over multiple backends
Credit @famarting

Adds support for situations whereby a single client is load balanced
across multiple instances of backends for the same durable task replica
set. This can be the case when there are multiple backends routable
behind a gateway which a durable task client routes to.

Backend RPCs have been added to for completion and wait signal state to
be moved to a set of backends.

Signed-off-by: joshvanl <me@joshvanl.dev>